refactor: move models package to internal/orm
