+++ nebula/components/home/sections/highlights_templ.go
@@ -8,7 +8,7 @@ package sections
 import "github.com/a-h/templ"
 import templruntime "github.com/a-h/templ/runtime"
 
-import "github.com/onsonr/sonr/nebula/models"
+import "github.com/onsonr/sonr/internal/orm/models"
 
 func Highlights(highlights *models.Highlights) templ.Component {
 	return templruntime.GeneratedTemplate(func(templ_7745c5c3_Input templruntime.GeneratedComponentInput) (templ_7745c5c3_Err error) {
